[CL] Extract create kinetic reactant.
